Update B2C token retrieval endpoint docs
As some people in de comments mention, the Token endpoint URL is incorrect. Also the token endpoint does not return an ```access_token``` but an ```id_token```, I'm not sure if this is an error in the docs or in the behaviour of the API but right now it's confusing.